requirements.txt
requests>=2.25.1
schedule>=1.1.0
web3>=5.28.0
python-telegram-bot>=13.7
python-dotenv>=0.19.0
